requirements.txt
numpy>=2.3.5
pandas>=2.3.3
scikit-learn>=1.7.2
joblib>=1.5.2
streamlit-nightly
